fix: add ModelInstanceTask and fix export path
