+++ packages/toolkit/src/lib/vdp-sdk/connector/index.ts
@@ -1,7 +1,3 @@
-import type { ConnectorState, Connector, ConnectorDefinition } from "./types";
-
-export type { ConnectorState, Connector, ConnectorDefinition };
-
 export * from "./source";
-
 export * from "./destination";
+export * from "./types";
